% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_display.R
\name{appFUN}
\alias{appFUN}
\title{A function to plot and map air quality data.
This function ... Written by Dr. Kyle Alberti (kyle.alberti@aeroqual.com).}
\usage{
appFUN(x, time.zone = "Pacific/Auckland", data.frequency = "60 min",
  dashboard.title = "AQY deployment", cols.to.display = c("O3", "PM2.5",
  "NO2"), site.col = "Serial", lat.col = "LAT", long.col = "LONG")
}
\arguments{
\item{x}{A csv file of the data. The data are set to wide format.}

\item{time.zone}{A string identifying the time zone for the data. Here, the default is `Pacific/Auckland`.}

\item{data.frequency}{A string identifying the averaging time of the data. Here, the default is `60 min`.}

\item{dashboard.title}{A string identifying the title that will appear on the dashboard. Here, `AQY deployment`.}

\item{cols.to.display}{A list of strings in x that are to be plotted. Here, the default is `c('O3","PM2.5","NO2")`.}

\item{site.col}{The name of the column in x that identifies the units or the sites. Here, the default is `Serial`.}

\item{lat.col}{Latitude column.}

\item{long.col}{Longitude column.}
}
\description{
A function to plot and map air quality data.
This function ... Written by Dr. Kyle Alberti (kyle.alberti@aeroqual.com).
}
\examples{
appFUN()
}
