% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksFUN.R
\name{ksFUN}
\alias{ksFUN}
\title{Running Kolmogorov-Smirnov (KS) Two-Sample Test.}
\usage{
ksFUN(x, obs, date, proxy, all.data = TRUE, theta = NA, tau = NA,
  window.length = 72)
}
\arguments{
\item{obs}{The data under investigation.}

\item{date}{A required column with the format 'YYYY-MM-DD HH:MM:SS'.}

\item{proxy}{The comparison data.}

\item{all.data}{This decides whether to use all provided data (all.data - `TRUE`), or to run the function on live data (`FALSE`), which makes the function possible to run as a schedule for real-time updating. Defaults is `TRUE`.}

\item{theta}{This sets the test threshold on whether to flag the data. Defaults to `NA` (no flags given).}

\item{tau}{This sets the day threshold on whether to flag the data given consistent theta flags. Defaults to `NA` (no flags given).}

\item{window.length}{This defines the length of the sampled window, defined by row number. Default is 72 (three days of hourly data).}
}
\description{
This function gives the running KS test for an observation, relative to a proxy from some other observation/s.
}
\examples{
ksFUN()
}
