% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMap.R
\name{loadMap}
\alias{loadMap}
\title{Mapping Function.}
\usage{
loadMap(x, multiple.files = FALSE, set.date, time.zone, obs, id, tier = NA,
  lat, lon, statistic = "latest", dest, leaflet.title = "Leaflet Map")
}
\arguments{
\item{x}{Either the file in R (default) or the location of a list of files to upload (if multuple.files is set to `TRUE`).}

\item{multiple.files}{An option if multiple files saved locally are to be plotted. Can be useful if there are a few daily files saved that need combining. This option is useful if the function is to run continuously.}

\item{set.date}{Latest date for the examined sample.}

\item{time.zone}{Time zone of the location (e.g. `Pacific/Auckland`).}

\item{obs}{Column under observation. Should be a numeric value. Can be multiple columns (e.g. "c("colA","colB")).}

\item{id}{Identifier column (e.g. site location name).}

\item{tier}{An additional identifier column, if needed (e.g. different measurement techniques).}

\item{lat}{Latitude.}

\item{lon}{Longitude.}

\item{statistic}{Choice of summary statistic. Options are `latest` (default), `mean`, `median`, `sd`, `min`, and `max`.}

\item{dest}{Location where the interactive map will be saved on the local drive.}

\item{leaflet.title}{A descriptor of the leaflet used in naming the HTML output.}
}
\description{
Function set up to work in near real-time, where (multiple) files can be created and saved to a working directory or data come from inside R environment. Summaries use the last seven days.
}
\examples{
loadMap()
}
