% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestNSitesFUN.R
\name{nearestNSitesFUN}
\alias{nearestNSitesFUN}
\title{Find the Nearest n Sites.}
\usage{
nearestNSitesFUN(x, id, lat, lon, n, weighted = F)
}
\arguments{
\item{id}{Name of the location (as character string).}

\item{lat}{Latitude column.}

\item{lon}{Longitude column.}

\item{n}{Number of nearest sites to find.}
}
\description{
A function for finding the closest n sites to a location.
}
\examples{
nearestNSitesFUN()
}
