% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestsiteFUN.R
\name{nearestsiteFUN}
\alias{nearestsiteFUN}
\title{This function gives the proxy type of nearest site data.}
\usage{
nearestsiteFUN(x, obs, group, lat, lon)
}
\arguments{
\item{obs}{This is the column being examined.}

\item{group}{This is the grouping variable.}

\item{lat}{Latitude column.}

\item{lon}{Longitude column.}
}
\description{
This function gives the proxy type of nearest site data.
}
\examples{
nearestsiteFUN()
}
