% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitesInNKMFUN.R
\name{sitesInNKMFUN}
\alias{sitesInNKMFUN}
\title{Find All Sites Within n Radius.}
\usage{
sitesInNKMFUN(x, id, lat, lon, n, as.list = F)
}
\arguments{
\item{id}{Name of the location (as character string).}

\item{lat}{Latitude column.}

\item{lon}{Longitude column.}

\item{n}{Number of nearest sites to find.}

\item{as.list}{Option to unite the selected sites into a string (can be useful for later selection).}
}
\description{
A function for finding all sites within a radius around a location.
}
\examples{
sitesInNKMFUN()
}
