% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMap.R
\name{loadMap}
\alias{loadMap}
\title{A mapping function. It is set up to work in a near real-time manner, where (multiple) daily files are created and saved to a location. Plots use the last seven days.}
\usage{
loadMap(x, multiple.files = FALSE, set.date = Sys.time(), obs, time.zone,
  id, tier = NA, lat, lon, dest, statistic = "latest",
  leaflet.title = "Leaflet Map")
}
\arguments{
\item{x}{This is either the file (default) or the location of a list of files to upload (if multuple.files is set to `TRUE`).}

\item{multiple.files}{An option if multiple files saved locally are to be plotted. Can be useful if there are a few daily files saved that need combining. This option is useful if the function is to run continuously.}

\item{set.date}{This is the latest date for the examined sample. Defaults to `Sys.time()`.}

\item{obs}{This is the column under observation. Should be a numeric value. Can be multiple columns (e.g. "c("colA","colB")).}

\item{time.zone}{This is the time zone of the location (e.g. `Pacific/Auckland`).}

\item{id}{This is an identifier column (e.g. site location name).}

\item{tier}{This is an additional identifier column, if needed (e.g. different measurement techniques).}

\item{lat}{Latitude.}

\item{lon}{Longitude.}

\item{dest}{This is the location where the interactive map will be saved on the local drive.}

\item{statistic}{This is the choice for statistic to be presented. Options are `latest` (default), `mean`, `median`, `sd`, `min`, and `max`.}

\item{leaflet.title}{This is some descriptor of the leaflet, if needed.}

\item{date}{This is the date column (set up as 'YYYY-MM-DD HH:MM:SS').}
}
\description{
This function gives a map with summary pop-ups for a set of observations.
}
\examples{
loadMap()
}
